👷 Add Dependabot configuration file
Introduce a new Dependabot configuration to the repository to enable automatic dependency updates on a weekly schedule. The configuration specifies the package ecosystem and the location of the package manifests, although the ecosystem is yet to be defined. Further adjustments may be needed to tailor the updates to the project's requirements.