Add .gitignore to exclude common directories
This commit introduces a .gitignore file in the .trunk directory. It
will prevent accidental inclusion of compiled output, logs, actions,
notifications, tools, and plugin directories, as well as specific
YAML configuration files, into the version control system. This is
important to keep the repository clean and to ensure that only
source files and necessary project files are tracked.