Add initial Trunk configuration file
This commit introduces the trunk.yaml file, which sets up the Trunk CLI
tool for our project. It specifies the version of the CLI, the plugins
source, the required runtimes, and the list of enabled linters. It also
configures certain actions to be disabled and enables the
trunk-upgrade-available action. This configuration will help maintain
code quality and enforce coding standards across the project.